add new events for renamed observed area
